handled delete category and tags

+++ app/Http/Controllers/TagsController.php
@@ -97,6 +97,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts()->count() > 0){
+            session()->flash('error','Tag Cannot be deleted beacuse this tag contains some posts');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success','Tag deleted successfully.');
         return redirect(route('tags.index'));
